rename savedday files and routes to create day
