+++ src/pages/CreateDay/index.js
@@ -1,0 +1,1 @@
+export { default as CreateDay } from './CreateDay.js';
