--- src/pages/SaveDay/index.js
@@ -1,1 +1,0 @@
-export { default as SaveDay } from './SaveDay.js';
